requirements.txt
aiohttp==3.0.1
amqp==2.2.2
async-timeout==2.0.0
attrs==17.4.0
beautifulsoup4==4.6.0
billiard==3.5.0.3
celery==4.1.0
certifi==2017.11.5
chardet==3.0.4
coverage==4.4.2
dj-database-url==0.4.2
Django==2.0
django-celery-results==1.0.1
django-cors-headers==2.0.2
django-debug-toolbar==1.9.1
django-extensions==1.9.8
django-jet==1.0.6
django-rest-framework==0.1.0
djangorestframework==3.7.2
djangorestframework-camel-case==0.2.0
djangorestframework-jwt==1.11.0
fake-useragent==0.1.10
gevent==1.2.2
greenlet==0.4.12
gunicorn==19.7.1
idna==2.6
idna-ssl==1.0.0
kombu==4.1.0
multidict==4.1.0
oauthlib==2.0.6
psycopg2==2.7.3.2
PyJWT==1.5.3
python-magic==0.4.15
pytz==2017.3
requests==2.18.4
selenium==3.9.0
six==1.11.0
sqlparse==0.2.4
typing==3.6.2
urllib3==1.22
vine==1.1.4
whitenoise==3.3.1
yarl==1.1.1
